Fixed some type conversions on readadc, added yawhold.h

+++ Firmware/QuadCopter/adc.h
@@ -57,7 +57,7 @@ void readADC(){
     }  
 }
 
-long getRawADC(int chnl){
+float getRawADC(int chnl){
     return adcValue[chnl]; 
 }
